rename discnet_lib to discnet
rename
