+++ discnet_test/CMakeLists.txt
@@ -34,7 +34,7 @@ target_link_libraries(discnet_test PRIVATE wsock32 ws2_32 IPHLPAPI wbemuuid bcry
 target_link_libraries(discnet_test PRIVATE Boost::log)
 target_link_libraries(discnet_test PRIVATE fmt::fmt)
 target_link_libraries(discnet_test PRIVATE GTest::gmock GTest::gtest GTest::gmock_main GTest::gtest_main)
-target_link_libraries(discnet_test PRIVATE discnet_lib)
+target_link_libraries(discnet_test PRIVATE discnet)
 target_link_libraries(discnet_test PRIVATE whatlog::whatlog)
 
 ############# force copy of dependencies to build folder #############
